refactor: 일부 hook 경로 변경 및 style 통일
