+++ src/components/RestaurantList/hooks/useRestaurantList.ts
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import mockData from '../../../data/mockData.json';
 import getSortingFilteredList from '../../../utils/getSortingFilteredList';
+import type { Selector } from '../../../hooks/useSelector';
 import type { Restaurant } from '../../RestaurantItem/type';
-import type { Selector } from '../../SelectorSection/hooks/useSelector';
 
 const mock = mockData.restaurantList as Restaurant[];
